Moved message records into their own files
