+++ src/BrowserHost.Tests/Fakes/FakeBrowserApis.cs
@@ -1,5 +1,4 @@
 ﻿using BrowserHost.Features.ActionContext.Tabs;
-using BrowserHost.Features.Settings;
 using BrowserHost.Features.TabPalette;
 using BrowserHost.Features.TabPalette.DomainCustomization;
 using BrowserHost.Features.TabPalette.FindText;
@@ -65,14 +64,4 @@ public class FakeDomainCustomizationBrowserApi() : DomainCustomizationBrowserApi
 
     public override void CallClientApi(string api, string? arguments = null) =>
         Invocations.Add(new(api, arguments));
-}
-
-public class FakeSettingsBrowserApi() : SettingsBrowserApi(() => null), IFakeBrowserApi
-{
-    public List<BrowserApiInvocation> Invocations { get; } = [];
-
-    // This is different from the other fake apis because the implementation forwards the arguments to
-    // the current tab browser CallClientApi method rather than calling CallClientApi directly.
-    public override void SettingsLoaded(SettingUiStateDto settings) =>
-        Invocations.Add(new("settingsLoaded", settings.ToJsonObject()));
 }
