+++ src/BrowserHost/Features/ActionContext/PinnedTabs/PinnedTabsBackendApi.cs
@@ -4,12 +4,6 @@ using BrowserHost.Utilities;
 
 namespace BrowserHost.Features.ActionContext.PinnedTabs;
 
-public record PinTabCommand(string TabId) : ICommand;
-public record UnpinTabCommand(string TabId) : ICommand;
-
-public record TabPinnedEvent(string TabId) : IEvent;
-public record TabUnpinnedEvent(string TabId) : IEvent;
-
 public class PinnedTabsBackendApi(PubSub pubSub) : BackendApi
 {
     public void UnpinTab(string tabId) =>
