+++ src/BrowserHost/Features/ActionContext/Tabs/TabListBackendApi.cs
@@ -7,18 +7,6 @@ using System.Linq;
 
 namespace BrowserHost.Features.ActionContext.Tabs;
 
-public record ActivateTabCommand(string TabId) : ICommand;
-public record CloseTabCommand(string TabId) : ICommand;
-public record ChangeTabsCommand(TabUiStateDto[] Tabs, int EphemeralTabStartIndex, FolderUiStateDto[] Folders) : ICommand;
-
-public record TabActivatedEvent(string TabId, TabBrowser? PreviousTab) : IEvent;
-public record TabDeactivatedEvent(string TabId) : IEvent;
-public record TabClosedEvent(string TabId, TabBrowser Tab) : IEvent;
-public record TabsChangedEvent(TabUiStateDto[] Tabs, int EphemeralTabStartIndex, FolderUiStateDto[] Folders) : IEvent;
-public record TabUrlLoadedSuccessfullyEvent(string TabId) : IEvent;
-public record TabFaviconUrlChangedEvent(string TabId, string? NewFaviconUrl) : IEvent;
-public record TabBrowserCreatedEvent(TabBrowser TabBrowser) : IEvent;
-
 public record TabUiStateDto(string Id, string Title, string? Favicon, bool IsActive, DateTimeOffset Created);
 public record FolderUiStateDto(string Id, string Name, int StartIndex, int EndIndex);
