+++ src/BrowserHost/Features/ActionContext/Workspaces/WorkspacesBackendApi.cs
@@ -5,18 +5,6 @@ using System;
 
 namespace BrowserHost.Features.ActionContext.Workspaces;
 
-public record ActivateWorkspaceCommand(string WorkspaceId) : ICommand;
-public record CreateWorkspaceCommand(string WorkspaceId, string Name, string Icon, string Color) : ICommand;
-public record UpdateWorkspaceCommand(string WorkspaceId, string Name, string Icon, string Color) : ICommand;
-public record DeleteWorkspaceCommand(string WorkspaceId) : ICommand;
-public record ExpireEphemeralTabsCommand(string[] TabIds) : ICommand;
-
-public record WorkspaceActivatedEvent(string WorkspaceId) : IEvent;
-public record WorkspaceCreatedEvent(string WorkspaceId, string Name, string Icon, string Color) : IEvent;
-public record WorkspaceUpdatedEvent(string WorkspaceId, string Name, string Icon, string Color) : IEvent;
-public record WorkspaceDeletedEvent(string WorkspaceId) : IEvent;
-public record EphemeralTabsExpiredEvent(string[] TabIds) : IEvent;
-
 public class WorkspacesBackendApi(PubSub pubSub) : BackendApi
 {
     public void ActivateWorkspace(string workspaceId) =>
