+++ src/BrowserHost/Features/DragDrop/Messages.cs
@@ -1,0 +1,6 @@
+using BrowserHost.Utilities;
+
+namespace BrowserHost.Features.DragDrop;
+
+public record OpenDroppedFilesCommand(string[] FilePaths) : ICommand;
+public record DroppedFilesOpenedEvent(string[] FilePaths) : IEvent;
