+++ src/BrowserHost/Features/Settings/SettingsBackendApi.cs
@@ -6,12 +6,6 @@ using System.Linq;
 
 namespace BrowserHost.Features.Settings;
 
-public record SaveSettingsCommand(SettingUiStateDto Settings) : ICommand;
-public record StartSsoFlowCommand(string TabId, string OriginalDomain, string OriginalUrl) : ICommand;
-
-public record SettingsSavedEvent(SettingUiStateDto Settings) : IEvent;
-public record SsoFlowStartedEvent(string TabId, string OriginalDomain, string OriginalUrl) : IEvent;
-
 public record SettingUiStateDto(string? UserAgent, string[] SsoEnabledDomains, bool AutoAddSsoDomains);
 
 public class SettingsBackendApi(PubSub pubSub, SettingsFeature settingsFeature) : BackendApi
