+++ src/BrowserHost/Features/TabPalette/DomainCustomization/DomainCustomizationBackendApi.cs
@@ -3,14 +3,6 @@ using BrowserHost.Utilities;
 
 namespace BrowserHost.Features.TabPalette.DomainCustomization;
 
-public record ChangeDomainCustomizationCommand(string Domain, bool CssEnabled) : ICommand;
-public record EditDomainCssCommand(string Domain) : ICommand;
-public record RemoveDomainCustomCssCommand(string Domain) : ICommand;
-
-public record DomainCustomizationChangedEvent(string Domain, bool CssEnabled) : IEvent;
-public record DomainCssEditRequestedEvent(string Domain) : IEvent;
-public record DomainCustomCssRemovedEvent(string Domain) : IEvent;
-
 public class DomainCustomizationBackendApi(PubSub pubSub) : BackendApi
 {
     public void SetCssEnabled(bool enabled)
