+++ src/BrowserHost/Features/TabPalette/Messages.cs
@@ -1,0 +1,9 @@
+using BrowserHost.Utilities;
+
+namespace BrowserHost.Features.TabPalette;
+
+public record RequestTabPaletteCommand() : Utilities.ICommand;
+public record TabPaletteRequestedEvent() : IEvent;
+
+public record DismissTabPaletteCommand() : Utilities.ICommand;
+public record TabPaletteDismissedEvent() : IEvent;
