+++ src/BrowserHost/Features/TabPalette/TabCustomization/TabCustomizationBackendApi.cs
@@ -3,12 +3,6 @@ using BrowserHost.Utilities;
 
 namespace BrowserHost.Features.TabPalette.TabCustomization;
 
-public record ChangeTabCustomTitleCommand(string TabId, string? CustomTitle) : ICommand;
-public record ChangeTabDisableFixedAddressCommand(string TabId, bool IsDisabled) : ICommand;
-
-public record TabCustomTitleChangedEvent(string TabId, string? CustomTitle) : IEvent;
-public record TabDisableFixedAddressChangedEvent(string TabId, bool IsDisabled) : IEvent;
-
 public class TabCustomizationBackendApi(PubSub pubSub) : BackendApi
 {
     public void SetCustomTitle(string? newTitle)
